draw: un-flatten the Vulkan rail, so neither rail owns `runtime::draw`'s namespace
The mirror of naming the Metal rail. `pub use vulkan::*` put sixty-odd items of
one rail directly into `runtime::draw`, so `runtime::draw::encode_draw_chain`,
`runtime::draw::forget_plane_draw_ring` and `runtime::draw::GvaSpan` were all
Vulkan's — and a reader could not tell from the path which of the two rails, or
neither, they had landed in.

* Forty-six `cfg(backend-vulkan)` items move from the shared module into
  `draw::vulkan`, and the glob re-export goes. They are that rail's working
  parts — the fragment-unbound scan, the m2v failure boundary, the engine state
  translations, the compositor plane ring, the type-11 memo — and they sat in
  the shared module only because the flat re-export left nowhere else for a
  `runtime::draw::<name>` caller to look.
* Callers outside the rail name it: `state`, `scanout`, `exec`, `compute_exec`,
  `writeback_debt` and `vk_engine_parity` now read `draw::vulkan::…`. Every one
  of those call sites was already inside a `cfg(backend-vulkan)` block; the path
  now says so too.
* The two hooks the shared request builders call into the rail —
  `publish_surface_store`'s plane-store note and `mrt_draw_request`'s attachment
  sample-count override — are spelled `vulkan::` at the call, which is what they
  always were.

The rail internals `draw::tests` reads are `pub(super)` rather than private:
that module holds both rails' tests, and forty Vulkan-rail tests are interleaved
with the neutral ones, so moving them the way the Metal rail's four moved is a
separate job. `pub(super)` is the accurate statement in the meantime — the
parent that owns both rails may see them, and nothing outside `runtime::draw`
can.

No behavior change on either arm. Metal 1315 pass, Vulkan 2173 pass, both
serial; feature-matrix green with per-arm test counts held.
