+++ crates/reims-vgpu/src/model/state.rs
@@ -3014,7 +3014,7 @@ impl DeviceState {
         // so it is dropped with the mapping: bounded by the live compositor
         // surfaces, and a recycled id cannot inherit a predecessor's passes.
         #[cfg(feature = "backend-vulkan")]
-        crate::runtime::draw::forget_plane_draw_ring(mapping_id);
+        crate::runtime::draw::vulkan::forget_plane_draw_ring(mapping_id);
         // Prune the dense-frame seq: a recycled mapping id must not inherit a
         // stale predecessor's dense seq.
         self.present.dense_frame_seq.remove(&mapping_id);
